refactor: export optional chaining transform
